commit 1

+++ src/main/java/com/example/travel/service/UserService.java
@@ -33,7 +33,7 @@ public interface UserService {
 
     Integer updateUserPayPasswordByOld(UpdateUserPayPasswordRequest updateUserPayPasswordRequest);
 
-    Integer investMoney(InvestMoney investMoney);
+    Double investMoney(InvestMoney investMoney);
 
     Double selectUserLastMoney(IdRequest idRequest);
